+++ Source/ACE/Managers/LandblockManager.cs
@@ -6,6 +6,7 @@ using ACE.Network;
 using ACE.Network.GameMessages.Messages;
 using log4net;
 using ACE.Common;
+using ACE.Entity.Events;
 
 namespace ACE.Managers
 {
@@ -44,10 +45,10 @@ namespace ACE.Managers
 
         // TODO: Need to be able to read the position of an object on the landblock and get information about that object CFS
 
-        public static void RemoveObject(WorldObject worldObject)
+        public static void RemoveObject(WorldObject worldObject, bool neighbors)
         {
-            Landblock block = GetLandblock(worldObject.Location.LandblockId, true);
-            block.RemoveWorldObject(worldObject.Guid, false);
+            Landblock block = GetLandblock(worldObject.Location.LandblockId, neighbors);
+            block.RemoveWorldObject(worldObject.Guid, neighbors);
         }
 
         /// <summary>
@@ -68,7 +69,24 @@ namespace ACE.Managers
             return block.GetWorldObject(targetId);
         }
 
-        /// <summary>
+        public static void BroadcastByLandblockID(BroadcastEventArgs args, bool propogate, Quadrant quadrant, LandblockId landBlockId)
+        {
+            Landblock block = GetLandblock(landBlockId, false);
+            block.Broadcast(args, propogate, quadrant);
+        }
+
+        public static Position GetWorldObjectPositionByLandblockID(ObjectGuid objectId, LandblockId landBlockId)
+        {
+            Landblock block = GetLandblock(landBlockId, false);
+            return block.GetWorldObjectPosition(objectId);
+        }
+
+        public static float GetWorldObjectEffectiveUseRadiusByLandblockID(ObjectGuid objectId, LandblockId landBlockId)
+        {
+            Landblock block = GetLandblock(landBlockId, false);
+            return block.GetWorldObjectEffectiveUseRadius(objectId);
+        }
+
         /// gets the landblock specified, creating it if it is not already loaded.  will create all
         /// adjacent landblocks if propagate is true (outdoor world roaming).
         /// </summary>