+++ Source/ACE/Network/GameAction/Actions/GameActionBuyItems.cs
@@ -3,6 +3,7 @@ using ACE.Database;
 using ACE.Entity;
 using ACE.Entity.Enum;
 using ACE.Managers;
+using ACE.Network.GameAction.QueuedGameActions;
 using ACE.Network.GameEvent;
 using ACE.Network.GameEvent.Events;
 using ACE.Network.GameMessages.Messages;
@@ -36,7 +37,7 @@ namespace ACE.Network.GameAction.Actions
             uint i_alternateCurrencyID = message.Payload.ReadUInt32();
 
             // todo: take into account other currencyIds other then assuming default
-            QueuedGameAction action = new QueuedGameAction(vendorID, items, GameActionType.Buy);
+            QueuedGameAction action = new QueuedGameActionBuy(vendorID, items);
             session.Player.AddToActionQueue(action);
             return;
         }