+++ Source/ACE/Network/GameAction/QueuedGameAction.cs
@@ -7,118 +7,52 @@ namespace ACE.Network.GameAction
 {
     public class QueuedGameAction
     {
-        public QueuedGameAction(uint objectId, GameActionType actionType)
-        {
-            this.ObjectId = objectId;
-            this.ActionType = actionType;
-            this.StartTime = WorldManager.PortalYearTicks;
-            this.EndTime = this.StartTime;
-        }
+        public void Handler(Player obj) { Handle(obj); }
+        protected virtual void Handle(Player obj) { }
 
-        public QueuedGameAction(uint objectId, uint secondaryObjectId, GameActionType actionType)
-        {
-            this.ObjectId = objectId;
-            this.SecondaryObjectId = secondaryObjectId;
-            this.ActionType = actionType;
-            this.StartTime = WorldManager.PortalYearTicks;
-            this.EndTime = this.StartTime;
-        }
+        public void Handler(Monster obj) { Handle(obj); }
+        protected virtual void Handle(Monster obj) { }
 
-        public QueuedGameAction(uint objectId, UniversalMotion motion, GameActionType actionType)
-        {
-            this.ObjectId = objectId;
-            this.Motion = motion;
-            this.ActionType = actionType;
-            this.StartTime = WorldManager.PortalYearTicks;
-            this.EndTime = this.StartTime;
-        }
+        public void Handler(Door obj) { Handle(obj); }
+        protected virtual void Handle(Door obj) { }
 
-        public QueuedGameAction(uint objectId, UniversalMotion motion, double duration, bool respectDelay, GameActionType actionType)
-        {
-            this.ObjectId = objectId;
-            this.Motion = motion;
-            this.RespectDelay = respectDelay;
-            this.ActionType = actionType;
-            this.StartTime = WorldManager.PortalYearTicks;
-            this.EndTime = this.StartTime + duration;
-        }
+        public void Handler(Container obj) { Handle(obj); }
+        protected virtual void Handle(Container obj) { }
 
-        public QueuedGameAction(uint objectId, WorldObject worldObject, bool respectDelay, GameActionType actionType)
-        {
-            this.ObjectId = objectId;
-            this.WorldObject = worldObject;
-            this.RespectDelay = respectDelay;
-            this.ActionType = actionType;
-            this.StartTime = WorldManager.PortalYearTicks;
-            this.EndTime = this.StartTime;
-        }
+        public void Handler(WorldObject obj) { Handle(obj); }
+        protected virtual void Handle(WorldObject obj) { }
 
-        public QueuedGameAction(uint objectId, WorldObject worldObject, bool respectDelay, bool onlyRemove, GameActionType actionType)
-        {
-            this.ObjectId = objectId;
-            this.WorldObject = worldObject;
-            this.RespectDelay = respectDelay;
-            this.OnlyRemove = onlyRemove;
-            this.ActionType = actionType;
-            this.StartTime = WorldManager.PortalYearTicks;
-            this.EndTime = this.StartTime;
-        }
+        public LandblockId LandBlockId { get; protected set; }
 
-        /// <summary>
-        /// Sends a player to a location
-        /// </summary>
-        /// <param name="objectId"></param>
-        /// <param name="newQueuedPosition">Teleport location</param>
-        public QueuedGameAction(uint objectId, Position newQueuedPosition, GameActionType actionType)
-        {
-            this.ObjectId = objectId;
-            this.ActionLocation = newQueuedPosition;
-            this.ActionType = actionType;
-        }
+        public uint ObjectId { get; protected set; }
 
-        /// <summary>
-        /// Send a player killed event
-        /// </summary>
-        /// <param name="objectId">Victim's Object Guid</param>
-        /// <param name="secondaryObjectId">Killer's Object Guid</param>
-        /// <param name="broadcastMessage">Text string for chat output, usually containing the death message</param>
-        public QueuedGameAction(string broadcastMessage, uint objectId, uint secondaryObjectId, GameActionType actionType)
-        {
-            this.ObjectId = objectId;
-            this.SecondaryObjectId = secondaryObjectId;
-            this.ActionBroadcastMessage = broadcastMessage;
-            this.ActionType = actionType;
-        }
+        public uint SecondaryObjectId { get; protected set; }
 
-        public QueuedGameAction(uint objectId, List<ItemProfile> profileItems, GameActionType actionType)
-        {
-            this.ObjectId = objectId;
-            this.ProfileItems = profileItems;
-            this.ActionType = actionType;
-        }
+        public WorldObject WorldObject { get; protected set; }
 
-        public uint ObjectId { get; private set; }
+        public string ActionBroadcastMessage { get; protected set; }
 
-        public uint SecondaryObjectId { get; private set; }
+        // this is weird, its only realy used now for teleport ?
+        // maybe it can go away soon enough.
+        public GameActionType ActionType { get; protected set; }
 
-        public WorldObject WorldObject { get; private set; }
+        public Position ActionLocation { get; protected set; }
 
-        public string ActionBroadcastMessage { get; private set; }
+        public UniversalMotion Motion { get; protected set; }
 
-        public GameActionType ActionType { get; private set; }
+        public bool OnlyRemove { get; protected set; }
 
-        public Position ActionLocation { get; private set; }
-
-        public UniversalMotion Motion { get; private set; }
-
-        public bool OnlyRemove { get; private set; }
-
-        public bool RespectDelay { get; private set; }
+        public bool RespectDelay { get; protected set; }
 
         public double StartTime { get; set; }
 
         public double EndTime { get; set; }
 
-        public List<ItemProfile> ProfileItems = new List<ItemProfile>();
+        public List<ItemProfile> ProfileItems { get; protected set; }
+
+        // Implement your own methods in your own QueuedGameAction class.
+        public QueuedGameAction()
+        {
+        }
     }
 }